chore: move integration test into integration folder
