+++ src/eth/evm/revm.rs
@@ -323,37 +323,3 @@ fn parse_revm_state(revm_state: RevmState, mut execution_changes: ExecutionChang
     Ok(execution_changes)
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*; use ethers_core::rand::thread_rng;
-    // Adjust this import to include your Revm and related structs
-    use fake::{Faker, Dummy};
-    use nonempty::nonempty;
-    use crate::config::StorageConfig;
-    use crate::eth::EthExecutor;
-    use crate::eth::primitives::{TransactionInput, Wei};
-    use crate::eth::storage::test_accounts;
-
-    #[tokio::test]
-    async fn regression_test_for_gas_handling() {
-        let storage: StorageConfig = "inmemory".parse().unwrap(); //XXX we need to use a real storage
-        let storage = storage.init().await.unwrap();
-        let mut rng = thread_rng();
-        let mut fake_transaction_input = TransactionInput::dummy_with_rng(&Faker, &mut rng);
-        fake_transaction_input.nonce = 0.into();
-        fake_transaction_input.gas_price = 0.into();
-        fake_transaction_input.gas = 0.into();
-
-        let accounts = test_accounts();
-        let _ = storage.enable_test_accounts(accounts.clone()).await.unwrap();
-
-        let address = accounts.last().unwrap().address.clone();
-        fake_transaction_input.from = address;
-        fake_transaction_input.value = Wei::from(0u64);
-
-        let revm = Box::new(Revm::new(Arc::clone(&storage)));
-        let executor = EthExecutor::new(nonempty![revm], Arc::clone(&storage));
-
-        executor.transact(fake_transaction_input).await.unwrap();
-    }
-}
