+++ src/eth/sync_parser.rs
@@ -35,7 +35,7 @@ mod tests {
 
     #[test]
     fn test_parse_receipts() {
-        let fixture_path = Path::new(env!("CARGO_MANIFEST_DIR")).join("test").join("fixtures").join("receipt_fixture.json");
+        let fixture_path = Path::new(env!("CARGO_MANIFEST_DIR")).join("tests").join("fixtures").join("receipt_fixture.json");
 
         // This JSON string represents a single transaction receipt. For testing, it needs to be wrapped in an array.
         let single_receipt_json = fs::read_to_string(fixture_path).expect("Failed to read fixture file");
@@ -52,7 +52,7 @@ mod tests {
 
     #[test]
     fn test_parse_block() {
-        let fixture_path = Path::new(env!("CARGO_MANIFEST_DIR")).join("test").join("fixtures").join("block_fixture.json");
+        let fixture_path = Path::new(env!("CARGO_MANIFEST_DIR")).join("tests").join("fixtures").join("block_fixture.json");
 
         // This JSON string represents a single transaction receipt. For testing, it needs to be wrapped in an array.
         let block_json = fs::read_to_string(fixture_path).expect("Failed to read fixture file");
