+++ tests/fixtures/receipt_fixture.json
